moving fuctions to gameComponest folder
