+++ gameComponents/newGame.py
@@ -1,0 +1,8 @@
+def new_game():
+    response = input("Press (Yes) to play again, Any key to quit\n")
+    response = response.upper()
+    if response == "YES":
+        return True
+    else:
+        return False
+
